Jitter packets that WERE broadcasted
Before, packets that were about to be broadcasted were jittered.
However this was conceptually wrong, as we wanted to reduce the number
of collisions caused by concurrent retransmissions. This is why it
is relevent if other nodes received the packet as well. The initial
packet, for example, does not need to be jittered (no one else will send
this packet at the same time)