Introduce continuous Merkle flow
The CastorContinuousMerkleFlow always maintains two Merkle flows. If the the current
Merkle flow dies, it migrates to using the next one. The dead flow is discarded and
a new flow is created as the 'next flow'.
The 'next flow' is needed to compute a commitment to the next flow (the hash of the next
Merkle tree root, i.e., the flow identifier).

The idea is to be able to reuse old flow state information at the intermediate nodes,
so to avoid the initial flooding phase. Currently, this state migration is not yet
implemented.