Safety commit
Code is currently broken. Note, we probably need to consider two
different kinds of NodeIds: one high layer (IP, which also defines the
flow), and one low layer (Ethernet, which is equivalent to a NeighborId)

So we should probably differentiate between EndPoint and Neighbor...

In some implementations such as Apple's MCF both could be equivalent.
In most other cases, however, we will be using IP and Ethernet.