Build Merkle tree based on nonce
Destination nodes build the Merkle tree based on a public nonce and a shared secret which are used as the input to a PRNG, specifically, a stream cipher.
The nonce is included in the first packets of a flow until the first ACK is received at the source. The advantage of this is, that the nonce (or previously encrypted ACK authenticator) does not need to be included in all PKTs, effectively reducing the header size. Also, it is no longer required for source and destination to perform encryption/decryption operations.

Currently, the nonce is still included in all PKTs, and SYN set in all PKTs.